Installation on Suse nodes is working

+++ services_setup/base-eskimo/install-eskimo-base-system.sh
@@ -64,11 +64,56 @@ if [[ -f "/etc/debian_version" ]]; then
         exit -101
     fi
 else
+    # works for both suse and RHEL
     if [[ ! `$pidof_command /usr/lib/systemd/systemd` ]]; then
         echoerr "Systemd is not running on node !"
         exit -101
     fi
 fi
+
+
+function enable_docker() {
+
+    echo "  - Enabling docker service"
+    sudo systemctl enable docker >>/tmp/install_docker 2>&1
+    if [[ $? != 0 ]]; then
+        echoerr "Unable to enable docker"
+        cat /tmp/install_docker 1>&2
+        exit -1
+    fi
+
+    echo "  - Starting docker service"
+    sudo systemctl start docker >>/tmp/install_docker 2>&1
+    if [[ $? != 0 ]]; then
+        echoerr "Unable to start docker"
+        cat /tmp/install_docker 1>&2
+        exit -1
+    fi
+
+    echo "  - Adding current user to docker group"
+    sudo usermod -a -G docker $USER >>/tmp/install_docker 2>&1
+    if [[ $? != 0 ]]; then
+        echoerr "Unable to add user $USER to docker"
+        cat /tmp/install_docker 1>&2
+        exit -1
+    fi
+
+}
+
+function install_docker_suse_based() {
+
+    rm -Rf /tmp/install_docker
+
+    sudo zypper install -y docker >>/tmp/install_docker 2>&1
+    if [[ $? != 0 ]]; then
+        echoerr "Unable to install required packages"
+        cat /tmp/install_docker 1>&2
+        exit -1
+    fi
+
+    enable_docker
+}
+
 
 
 function install_docker_redhat_based() {
@@ -109,29 +154,7 @@ function install_docker_redhat_based() {
         exit -1
     fi
 
-    echo "  - Enabling docker service"
-    sudo systemctl enable docker >>/tmp/install_docker 2>&1
-    if [[ $? != 0 ]]; then
-        echoerr "Unable to enable docker"
-        cat /tmp/install_docker 1>&2
-        exit -1
-    fi
-
-    echo "  - Starting docker service"
-    sudo systemctl start docker >>/tmp/install_docker 2>&1
-    if [[ $? != 0 ]]; then
-        echoerr "Unable to start docker"
-        cat /tmp/install_docker 1>&2
-        exit -1
-    fi
-
-    echo "  - Adding current user to docker group"
-    sudo usermod -a -G docker $USER >>/tmp/install_docker 2>&1
-    if [[ $? != 0 ]]; then
-        echoerr "Unable to add user $USER to docker"
-        cat /tmp/install_docker 1>&2
-        exit -1
-    fi
+    enable_docker
 }
 
 function install_docker_debian_based() {
@@ -198,29 +221,19 @@ function install_docker_debian_based() {
         exit -1
     fi
 
-    echo "  - Enabling docker service"
-    sudo systemctl enable docker >>/tmp/install_docker 2>&1
-    if [[ $? != 0 ]]; then
-        echoerr "Unable to enable docker"
-        cat /tmp/install_docker 1>&2
+    enable_docker
+}
+
+function install_suse_mesos_dependencies() {
+
+    echo " - Installing other Mesos dependencies"
+    sudo zypper install -y zlib-devel libcurl-devel openssl-devel cyrus-sasl-devel cyrus-sasl-plain cyrus-sasl-crammd5 apr-devel subversion-devel apr-util-devel >> /tmp/setup_log 2>&1
+     if [[ $? != 0 ]]; then
+        echoerr "Unable to install mesos dependencies"
+        cat /tmp/setup_log 1>&2
         exit -1
     fi
 
-    echo "  - Starting docker service"
-    sudo systemctl start docker >>/tmp/install_docker 2>&1
-    if [[ $? != 0 ]]; then
-        echoerr "Unable to start docker"
-        cat /tmp/install_docker 1>&2
-        exit -1
-    fi
-
-    echo "  - Adding current user to docker group"
-    sudo usermod -a -G docker $USER >>/tmp/install_docker 2>&1
-    if [[ $? != 0 ]]; then
-        echoerr "Unable to add user $USER to docker"
-        cat /tmp/install_docker 1>&2
-        exit -1
-    fi
 }
 
 function install_redhat_mesos_dependencies() {
@@ -267,8 +280,14 @@ if [[ $? != 0 ]]; then
     echo "  - docker is not installed. attempting installation"
     if [[ -f "/etc/debian_version" ]]; then
         install_docker_debian_based
-    else
+
+    elif [[ -f "/etc/redhat-release" ]]; then
+
         install_docker_redhat_based
+    else
+
+        # assuming suse
+        install_docker_suse_based
     fi
 fi
 
@@ -280,7 +299,7 @@ if [[ -f "/etc/debian_version" ]]; then
     # ignore this one if it fails
     sudo DEBIAN_FRONTEND=noninteractive apt-get -yq install attr >> /tmp/setup_log 2>&1
 
-else
+elif [[ -f "/etc/redhat-release" ]]; then
     sudo yum install -y net-tools anacron >> /tmp/setup_log 2>&1
     fail_if_error $? "/tmp/setup_log" -1
 
@@ -291,14 +310,28 @@ else
     sudo systemctl start crond >> /tmp/setup_log 2>&1
     fail_if_error $? "/tmp/setup_log" -1
 
+else
+
+    # assuming suse
+    sudo zypper install -y net-tools cron >> /tmp/setup_log 2>&1
+    fail_if_error $? "/tmp/setup_log" -1
+
+    echo "  - enabling cron"
+    sudo systemctl enable cron >> /tmp/setup_log 2>&1
+    fail_if_error $? "/tmp/setup_log" -1
+
+    sudo systemctl start cron >> /tmp/setup_log 2>&1
+    fail_if_error $? "/tmp/setup_log" -1
 fi
 
 
 echo "  - installing mesos dependencies"
 if [[ -f "/etc/debian_version" ]]; then
     install_debian_mesos_dependencies
-else
+elif [[ -f "/etc/redhat-release" ]]; then
     install_redhat_mesos_dependencies
+else
+    install_suse_mesos_dependencies
 fi
 
 
@@ -306,9 +339,15 @@ echo " - Installing gluster client"
 if [[ -f "/etc/debian_version" ]]; then
     sudo apt-get -y install glusterfs-client >> /tmp/setup_log 2>&1
     fail_if_error $? "/tmp/setup_log" -1
-else
+
+elif [[ -f "/etc/redhat-release" ]]; then
     sudo yum -y install glusterfs glusterfs-fuse >> /tmp/setup_log 2>&1
     fail_if_error $? "/tmp/setup_log" -1
+
+else
+    sudo zypper install -y glusterfs  >> /tmp/setup_log 2>&1
+    fail_if_error $? "/tmp/setup_log" -1
+
 fi
